extract popover to own file
